[expo-go] Remove bottom sheet (reanimated)
